# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shan Loray Beauty E-Commerce API",
        "description": "Complete REST API for the Shan Loray luxury beauty e-commerce platform",
        "contact": {
            "name": "Shan Loray API Support",
            "email": "api@shanloray.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://shanloray.com/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products with filtering and sorting",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "bestselling",
                                "newest",
                                "price_low",
                                "price_high",
                                "rating"
                            ]
                        }
                    },
                    {
                        "name": "badge",
                        "in": "query",
                        "description": "Filter by badge",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "bestseller",
                                "new",
                                "limited",
                                "signature"
                            ]
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "description": "Featured only",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products",
                "operationId": "a3211e37d25c399811b6118eb88e49af",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (min 2 chars)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "query": {
                                            "type": "string"
                                        },
                                        "count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a single product by slug",
                "operationId": "5efa959b82f6bb80e85a8c64320dd543",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Hydrating Face Serum"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hydrating-face-serum"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 149
                    },
                    "original_price": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "example": "AED"
                    },
                    "image": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.8
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 127
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 56
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}